pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zero-dex-research"
version = "0.1.0"
description = "Research framework for adversarial dark-pool routing with GNN-based RL agents"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "FLock.io"},
]
keywords = [
    "reinforcement-learning",
    "graph-neural-networks",
    "dark-pool",
    "byzantine-fault-tolerance",
    "zero-knowledge-proofs",
    "decentralized-exchange",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gymnasium>=1.0",
    "numpy>=1.26",
    "torch>=2.2",
    "torch-geometric>=2.5",
    "pyyaml>=6.0",
    "tensorboard>=2.16",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]
zk = [
    "py_ecc>=7.0",
]

[project.scripts]
zero-dex-experiment = "experiments.run_experiment:main"

[tool.setuptools.packages.find]
include = ["envs*", "agents*", "simulation*", "zk*", "experiments*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
